clean init
